Added New Command: GetGraphicsConfig
Allows access to the base graphics renderer in Android OPENGLES thru pointers for use with OpenXR.  Added same commands to OpenGL2 and Vulcan but not yet implemented.  This command has not been tested if it would affect successful compile on windows mac or linux.  Should not be pushed to main AGK yet.