Rotation Drift Fix Added
Fixed the bug caused by using rotateX/rotateY/rotateZ.  This bug only appeared in the meta quest version of openxr, and didn't happen in the steam version of openxr.  Properly correcting the issue for the meta would break the steam version.  By adding "m_Offset" to track for position drift, this works in both versions.